feat(api): Allow metacharacters to be escaped in event/issue searching (SEN-153)
`fnmatch.translate` doesn't currently support escaping metacharacters. Found some discussion about
this and a simple patch to allow escaping.

As part of testing that this works as expected, also found that `positionCaseInsensitive` only
accepts a needle (basically a string) to search for. This means that wildcard matching is doesn't
work for message in both issue and event search. Converted it over to use `match`, and retained
case insensitivity by using (?i) to set the case insensitive flag. Also added this to matching for
other values.

Also found that the handling of `message` in postgres also did not support regex matching. Switched
over to using it for wildcard matches, but only when using postgres. Use the normal `contains`
matching for non-wild card as usual.

I also added a hack in to both postgres and snuba to strip the `^$` characters from the
beginning/end of the regex, since I assume it makes sense to search for the regex all throughout
the message. Not sure if this is required for other fields, if not we could just remove it from
`translate` in general.